Cleanup of buffers and access interfaces
