+++ src/main/java/org/spongepowered/api/util/gen/MutableBiomeArea.java
@@ -30,7 +30,7 @@ import org.spongepowered.api.world.biome.BiomeType;
  * A mutable buffer for {@link BiomeType} data. This buffer has no direct relation
  * to the world and changes to it are not synchronized to the world.
  */
-public interface MutableBiomeArea extends BiomeArea {
+public interface MutableBiomeArea extends BiomeBuffer {
 
     /**
      * Sets the biome in the buffer at the given position.
